+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Types of changes:
 [Unreleased]
 
 - [It should be possible to exclude (temporary) database jobs while generating DDL.](https://github.com/paulissoft/oracle-tools/issues/22)
+- [Add SQL Datamodeler custom library and transformations scripts.](https://github.com/paulissoft/oracle-tools/issues/46)
 
 ## [1.5.0] - 2022-08-03